[refactor] 패키지명 변경 및 테이블명 오타 정정
