+++ src/main/java/com/ssafy/live/domain/recommendation/dto/RecommendationRequestDTO.java
@@ -1,4 +1,4 @@
-package com.ssafy.live.domain.recommdation.dto;
+package com.ssafy.live.domain.recommendation.dto;
 
 import lombok.Data;
 import lombok.Getter;
